+++ nequip/data/AtomicData.py
@@ -5,7 +5,7 @@ Authors: Albert Musaelian
 
 import warnings
 from copy import deepcopy
-from typing import Union, Tuple, Dict, Optional, List
+from typing import Union, Tuple, Dict, Optional, List, Set, Sequence
 from collections.abc import Mapping
 
 import numpy as np
@@ -22,6 +22,72 @@ from nequip.utils.torch_geometric import Data
 
 # A type representing ASE-style periodic boundary condtions, which can be partial (the tuple case)
 PBC = Union[bool, Tuple[bool, bool, bool]]
+
+_DEFAULT_NODE_FIELDS: Set[str] = {
+    AtomicDataDict.POSITIONS_KEY,
+    AtomicDataDict.WEIGHTS_KEY,
+    AtomicDataDict.NODE_FEATURES_KEY,
+    AtomicDataDict.NODE_ATTRS_KEY,
+    AtomicDataDict.ATOMIC_NUMBERS_KEY,
+    AtomicDataDict.ATOM_TYPE_KEY,
+    AtomicDataDict.FORCE_KEY,
+    AtomicDataDict.PER_ATOM_ENERGY_KEY,
+    AtomicDataDict.BATCH_KEY,
+}
+_DEFAULT_EDGE_FIELDS: Set[str] = {
+    AtomicDataDict.EDGE_CELL_SHIFT_KEY,
+    AtomicDataDict.EDGE_VECTORS_KEY,
+    AtomicDataDict.EDGE_LENGTH_KEY,
+    AtomicDataDict.EDGE_ATTRS_KEY,
+    AtomicDataDict.EDGE_EMBEDDING_KEY,
+}
+_DEFAULT_GRAPH_FIELDS: Set[str] = {
+    AtomicDataDict.TOTAL_ENERGY_KEY,
+}
+_NODE_FIELDS: Set[str] = set(_DEFAULT_NODE_FIELDS)
+_EDGE_FIELDS: Set[str] = set(_DEFAULT_EDGE_FIELDS)
+_GRAPH_FIELDS: Set[str] = set(_DEFAULT_GRAPH_FIELDS)
+
+
+def register_fields(
+    node_fields: Sequence[str] = [],
+    edge_fields: Sequence[str] = [],
+    graph_fields: Sequence[str] = [],
+) -> None:
+    r"""Register fields as being per-atom, per-edge, or per-frame.
+
+    Args:
+        node_permute_fields: fields that are equivariant to node permutations.
+        edge_permute_fields: fields that are equivariant to edge permutations.
+    """
+    node_fields: set = set(node_fields)
+    edge_fields: set = set(edge_fields)
+    graph_fields: set = set(graph_fields)
+    allfields = node_fields.union(edge_fields, graph_fields)
+    assert len(allfields) == len(node_fields) + len(edge_fields) + len(graph_fields)
+    assert (_NODE_FIELDS.union(_EDGE_FIELDS, _GRAPH_FIELDS)).isdisjoint(
+        allfields
+    ), "Cannot reregister a field that has already been registered"
+    _NODE_FIELDS.update(node_fields)
+    _EDGE_FIELDS.update(edge_fields)
+    _GRAPH_FIELDS.update(graph_fields)
+
+
+def deregister_fields(*fields: Sequence[str]) -> None:
+    r"""Deregister a field registered with ``register_fields``.
+
+    Silently ignores fields that were never registered to begin with.
+
+    Args:
+        *fields: fields to deregister.
+    """
+    for f in fields:
+        assert f not in _DEFAULT_NODE_FIELDS, "Cannot deregister built-in field"
+        assert f not in _DEFAULT_EDGE_FIELDS, "Cannot deregister built-in field"
+        assert f not in _DEFAULT_GRAPH_FIELDS, "Cannot deregister built-in field"
+        _NODE_FIELDS.discard(f)
+        _EDGE_FIELDS.discard(f)
+        _GRAPH_FIELDS.discard(f)
 
 
 class AtomicData(Data):