+++ tests/unit/data/test_dataset.py
@@ -1,3 +1,4 @@
+from attr import fields
 import numpy as np
 import pytest
 import tempfile
@@ -5,8 +6,9 @@ import torch
 
 from os.path import isdir, isfile
 
-import ase.data
+from ase.data import chemical_symbols
 from ase.io import write
+from torch.functional import _return_counts
 
 from nequip.data import (
     AtomicDataDict,
@@ -15,6 +17,7 @@ from nequip.data import (
     ASEDataset,
     dataset_from_config,
 )
+from nequip.data.transforms import TypeMapper
 from nequip.utils import Config
 
 
@@ -26,29 +29,30 @@ def ase_file(molecules):
         yield fp.name
 
 
-MAX_ATOMIC_NUMBER: int = 8
+MAX_ATOMIC_NUMBER: int = 5
+NATOMS = 3
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def npz():
-    natoms = 3
-    nframes = 4
+    natoms = NATOMS
+    nframes = 8
     yield dict(
         positions=np.random.random((nframes, natoms, 3)),
         force=np.random.random((nframes, natoms, 3)),
-        energy=np.random.random(nframes),
+        energy=np.random.random(nframes) * -600,
         Z=np.random.randint(1, MAX_ATOMIC_NUMBER, size=(nframes, natoms)),
     )
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def npz_data(npz):
     with tempfile.NamedTemporaryFile(suffix=".npz") as path:
         np.savez(path.name, **npz)
         yield path.name
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def npz_dataset(npz_data, temp_data):
     a = NpzDataset(
         file_name=npz_data,
@@ -67,7 +71,7 @@ def root():
 class TestInit:
     def test_init(self):
         with pytest.raises(NotImplementedError) as excinfo:
-            a = AtomicInMemoryDataset(root=None)
+            AtomicInMemoryDataset(root=None)
         assert str(excinfo.value) == ""
 
     def test_npz(self, npz_data, root):
@@ -106,7 +110,8 @@ class TestStatistics:
     def test_statistics(self, npz_dataset, npz):
 
         (eng_mean, eng_std), (Z_unique, Z_count) = npz_dataset.statistics(
-            [AtomicDataDict.TOTAL_ENERGY_KEY, AtomicDataDict.ATOMIC_NUMBERS_KEY]
+            fields=[AtomicDataDict.TOTAL_ENERGY_KEY, AtomicDataDict.ATOMIC_NUMBERS_KEY],
+            modes=["mean_std", "count"],
         )
 
         eng = npz["energy"]
@@ -130,7 +135,6 @@ class TestStatistics:
             [AtomicDataDict.ATOMIC_NUMBERS_KEY, AtomicDataDict.FORCE_KEY],
             modes=["count", "rms"],
         )
-        print("npz", npz["Z"])
 
         uniq, count = np.unique(npz["Z"][0].ravel(), return_counts=True)
         assert np.all(Z_unique.numpy() == uniq)
@@ -139,6 +143,110 @@ class TestStatistics:
         assert np.allclose(
             force_rms.numpy(), np.sqrt(np.mean(np.square(npz["force"][0])))
         )
+
+    def test_atom_types(self, npz_dataset):
+        ((avg_num_neigh, _),) = npz_dataset.statistics(
+            fields=[
+                lambda data: (
+                    torch.unique(
+                        data[AtomicDataDict.EDGE_INDEX_KEY][0], return_counts=True
+                    )[1],
+                    "node",
+                )
+            ],
+            modes=["mean_std"],
+        )
+        # They are all homogenous in this dataset:
+        assert (
+            avg_num_neigh
+            == torch.bincount(npz_dataset[0][AtomicDataDict.EDGE_INDEX_KEY][0])[0]
+        )
+
+    def test_edgewise_stats(self, npz_dataset):
+        ((avg_edge_length, std_edge_len),) = npz_dataset.statistics(
+            fields=[
+                lambda data: (
+                    (
+                        data[AtomicDataDict.POSITIONS_KEY][
+                            data[AtomicDataDict.EDGE_INDEX_KEY][1]
+                        ]
+                        - data[AtomicDataDict.POSITIONS_KEY][
+                            data[AtomicDataDict.EDGE_INDEX_KEY][0]
+                        ]
+                    ).norm(dim=-1),
+                    "edge",
+                )
+            ],
+            modes=["mean_std"],
+        )
+        # TODO: check correct
+
+
+class TestPerSpeciesStatistics:
+    @pytest.mark.parametrize("fixed_field", [True, False])
+    @pytest.mark.parametrize("mode", ["mean_std", "rms"])
+    def test_per_node_field(self, npz_dataset, fixed_field, mode):
+        # set up the transformer
+        npz_dataset = set_up_transformer(npz_dataset, not fixed_field, fixed_field)
+
+        (result,) = npz_dataset.statistics(
+            [AtomicDataDict.BATCH_KEY],
+            modes=[f"per_species_{mode}"],
+        )
+        print(result)
+
+    @pytest.mark.parametrize("alpha", [1e-10, 1e-6, 0.1, 0.5, 1])
+    @pytest.mark.parametrize("fixed_field", [True, False])
+    @pytest.mark.parametrize("full_rank", [True, False])
+    def test_per_graph_field(self, npz_dataset, alpha, fixed_field, full_rank):
+
+        npz_dataset = set_up_transformer(npz_dataset, full_rank, fixed_field)
+        if npz_dataset is None:
+            return
+
+        # get species count per graph
+        Ns = []
+        for i in range(npz_dataset.len()):
+            Ns.append(torch.bincount(npz_dataset[i][AtomicDataDict.ATOM_TYPE_KEY]))
+        n_spec = max(len(e) for e in Ns)
+        N = torch.zeros(len(Ns), n_spec)
+        for i in range(len(Ns)):
+            N[i, : len(Ns[i])] = Ns[i]
+        del n_spec
+        del Ns
+
+        if alpha == 1e-10:
+            ref_mean, ref_std, E = generate_E(N, 100, 0.0)
+        else:
+            ref_mean, ref_std, E = generate_E(N, 100, 0.5)
+
+        npz_dataset.data[AtomicDataDict.TOTAL_ENERGY_KEY] = E
+
+        ref_res2 = torch.square(
+            torch.matmul(N, ref_mean.reshape([-1, 1])) - E.reshape([-1, 1])
+        ).sum()
+
+        ((mean, std),) = npz_dataset.statistics(
+            [AtomicDataDict.TOTAL_ENERGY_KEY],
+            modes=["per_species_mean_std"],
+            kwargs={
+                AtomicDataDict.TOTAL_ENERGY_KEY
+                + "per_species_mean_std": {"alpha": alpha}
+            },
+        )
+
+        res = torch.matmul(N, mean.reshape([-1, 1])) - E.reshape([-1, 1])
+        res2 = torch.sum(torch.square(res))
+        print("residue", alpha, res2 - ref_res2)
+        print("mean", mean, ref_mean)
+        print("diff in mean", mean - ref_mean)
+        print("std", std, ref_std)
+
+        if alpha == 1e-10 and full_rank:
+            assert torch.allclose(mean, ref_mean, rtol=1e-1)
+            assert torch.allclose(std, torch.zeros_like(ref_mean), atol=1e-2)
+        # else:
+        #     assert res2 > ref_res2
 
 
 class TestReload:
@@ -181,8 +289,7 @@ class TestFromConfig:
                 file_name=npz_data,
                 root=root,
                 chemical_symbol_to_type={
-                    ase.data.chemical_symbols[an]: an - 1
-                    for an in range(1, MAX_ATOMIC_NUMBER)
+                    chemical_symbols[an]: an - 1 for an in range(1, MAX_ATOMIC_NUMBER)
                 },
                 **args,
             )
@@ -230,3 +337,54 @@ class TestFromList:
             assert np.array_equal(
                 mol.get_atomic_numbers(), dataset[i].to_ase().get_atomic_numbers()
             )
+
+
+def generate_E(N, mean, std):
+    torch.manual_seed(0)
+    ref_mean = torch.rand((N.shape[1])) * mean
+    t_mean = torch.ones((N.shape[0], 1)) * ref_mean.reshape([1, -1])
+    ref_std = torch.rand((N.shape[1])) * std
+    t_std = torch.ones((N.shape[0], 1)) * ref_std.reshape([1, -1])
+    E = torch.normal(t_mean, t_std)
+    return ref_mean, ref_std, (N * E).sum(axis=-1)
+
+
+def set_up_transformer(npz_dataset, full_rank, fixed_field):
+
+    if full_rank:
+
+        if fixed_field:
+            return
+
+        unique = torch.unique(npz_dataset.data[AtomicDataDict.ATOMIC_NUMBERS_KEY])
+        npz_dataset.transform = TypeMapper(
+            chemical_symbol_to_type={
+                chemical_symbols[n]: i for i, n in enumerate(unique)
+            }
+        )
+    else:
+        ntype = 2
+
+        # let all atoms to be the same type distribution
+        num_nodes = npz_dataset.data[AtomicDataDict.BATCH_KEY].shape[0]
+        if fixed_field:
+            del npz_dataset.data[AtomicDataDict.ATOMIC_NUMBERS_KEY]
+            del npz_dataset.data.__slices__[
+                AtomicDataDict.ATOMIC_NUMBERS_KEY
+            ]  # remove batch metadata for the key
+            new_n = torch.ones(NATOMS, dtype=torch.int64)
+            new_n[0] += ntype
+            npz_dataset.fixed_fields[AtomicDataDict.ATOMIC_NUMBERS_KEY] = new_n
+        else:
+            npz_dataset.fixed_fields.pop(AtomicDataDict.ATOMIC_NUMBERS_KEY, None)
+            new_n = torch.ones(num_nodes, dtype=torch.int64)
+            new_n[::NATOMS] += ntype
+            npz_dataset.data[AtomicDataDict.ATOMIC_NUMBERS_KEY] = new_n
+
+        # set up the transformer
+        npz_dataset.transform = TypeMapper(
+            chemical_symbol_to_type={
+                chemical_symbols[n]: i for i, n in enumerate([1, ntype + 1])
+            }
+        )
+    return npz_dataset